chore: refactor (moved duneprovider to context)
