+++ src/app/globals.css
@@ -3,11 +3,6 @@
 
 @custom-variant dark (&:is(.dark *));
 
-@theme {
-  --font-geist-mono: "Geist Mono", "sans-serif";
-  --font-geist: "Geist", "sans-serif";
-}
-
 @theme inline {
   --color-background: var(--background);
   --color-foreground: var(--foreground);
