+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Provider from "@/context/Provider";
+import Provider from "@/context/provider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
